==> scRNA-seq/3-Combination_Tumor_Yerly_Ganier.Rmd <==
---
title: "Combination tumor cells from Yerly + Ganier"
author: "Massimo Andreatta & Laura Yerly"
output: html_document
date: "2024-02-28"
---

```{r setup, include=FALSE,fig.width=16,fig.height=12}
renv::restore()
library(Seurat)
library(ggplot2)
```

# Set paths 

These .rds objects were prepared using the 1.* and 2.* scripts.
```{r}
Tumor_LY <- readRDS("cache/Tumor_LY.rds")
Tumor_CG <- readRDS("cache/Tumor_CG.rds")
```

# Merge Tumor cells, select samples with enough cancer cells
```{r}
Tumor.combined <- merge(x = Tumor_LY, y = Tumor_CG)
#Tumor.combined <- JoinLayers(Tumor.combined)
tab <- table(Tumor.combined$Sample)
tab 

pass <- names(tab)[tab > 100]
Tumor.combined <- subset(Tumor.combined, subset=Sample %in% pass)
```

## Set up some helper functions
```{r}
get_means_signature<- function(object,signature){
  means <- sapply(levels(object$seurat_clusters), function(x) {
    mean(object@meta.data[object$seurat_clusters == x, signature])
  })
  means
}

get_medians_signature<- function(object,signature){
  medians <- sapply(levels(object$seurat_clusters), function(x) {
    median(object@meta.data[object$seurat_clusters == x, signature])
  })
  medians
}
```


```{r fig.width=12}
Tumor.combined <- Tumor.combined |> FindVariableFeatures(nfeatures=2000) |>
  ScaleData()

Tumor.combined <- Tumor.combined |> RunPCA(npcs=30) |> RunUMAP(reduction = "pca", dims = 1:30)

a <- DimPlot(Tumor.combined, reduction = "umap", group.by = "Sample") + theme(aspect.ratio = 1)
b <- DimPlot(Tumor.combined, reduction = "umap", group.by = "annotation") + theme(aspect.ratio = 1)

a | b
```


View signatures
```{r fig.height=12, fig.width=20}
feats <- c("BCC","NormKerat","TCells","Melanocytes","BCells","Pericytes","Fibroblasts","MyeloidCells")
FeaturePlot(Tumor.combined, features=feats) & theme(aspect.ratio = 1,
  axis.title = element_blank(),
        axis.text = element_blank(),
        axis.ticks = element_blank())
```


## Clustering

```{r}
Tumor.combined <- FindNeighbors(object = Tumor.combined, dims=1:30)
Tumor.combined <- FindClusters(object = Tumor.combined, resolution = 2)
DimPlot(Tumor.combined, reduction = "umap", group.by = "seurat_clusters") + theme(aspect.ratio = 1)
```

```{r}
VlnPlot(Tumor.combined,features=feats,
        group.by = "seurat_clusters", stack = T,flip = T,assay = "RNA")+ NoLegend()
```


## Get mean expression by cluster
```{r}
BCC_means<-get_means_signature(Tumor.combined,"BCC")
NormKerat_means<-get_means_signature(Tumor.combined,"NormKerat")
TCells_means<-get_means_signature(Tumor.combined,"TCells")
Myeloid_means <- get_means_signature(Tumor.combined,"MyeloidCells")
Mast_means <- get_means_signature(Tumor.combined,"MastCells")
BCells_means <- get_means_signature(Tumor.combined,"BCells")
```


Any doublets?
```{r}
plot(x=BCC_means, y=TCells_means)
plot(x=BCC_means, y=Myeloid_means)
plot(x=BCC_means, y=BCells_means)
plot(x=BCC_means, y=Mast_means)
```

Remove doublets (more conservative, to make sure we remove all contaminants)
```{r}
is.doublet <- TCells_means > 0.05 | BCells_means > 0.1 | Mast_means > 0.2 | Myeloid_means > 0.1
pass <- names(is.doublet)[!is.doublet]

Tumor.combined <- subset(Tumor.combined, subset=seurat_clusters %in% pass)
```

```{r fig.width=12}
Tumor.combined <- Tumor.combined |> FindVariableFeatures(nfeatures=2000) |>
  ScaleData()

Tumor.combined <- Tumor.combined |> RunPCA(npcs=30) |> RunUMAP(reduction = "pca", dims = 1:30)

a <- DimPlot(Tumor.combined, reduction = "umap", group.by = "Sample") + theme(aspect.ratio = 1)
b <- DimPlot(Tumor.combined, reduction = "umap", group.by = "annotation") + theme(aspect.ratio = 1)

a | b
```

View signatures
```{r fig.height=12, fig.width=20}
feats <- c("BCC","NormKerat","TCells","Melanocytes","BCells","Pericytes","Fibroblasts","MyeloidCells")
FeaturePlot(Tumor.combined, features=feats) & theme(aspect.ratio = 1,
  axis.title = element_blank(),
        axis.text = element_blank(),
        axis.ticks = element_blank())
```

By sample, as an extra check
```{r}
seu.list <- SplitObject(Tumor.combined, split.by = "Sample")

seu.list <- lapply(seu.list, function(x) {
  x <- NormalizeData(x) |> FindVariableFeatures(nfeatures=2000, verbose = F) |> 
    ScaleData() |> RunPCA(npcs=30, verbose = F) |> RunUMAP(reduction = "pca", dims = 1:30, verbose = F)
  
  x <- FindNeighbors(x, dims=1:30, verbose=F) |> FindClusters(resolution = 1, verbose = F)
  x
})
```

```{r fig.height=12, fig.width=20}
pll <- lapply(seu.list, function(x) {
  FeaturePlot(x, features=feats) & theme(aspect.ratio = 1,
                                                       axis.title = element_blank(),
                                                       axis.text = element_blank(),
                                                       axis.ticks = element_blank())
})
pll
```


# Save 
```{r}
tab <- table(Tumor.combined$Sample)
as.data.frame(tab)

file <- "cache/Tumor_combination_LY_CG.rds"
saveRDS(Tumor.combined, file)

```
