==> scRNA-seq/P3-MPs_vs_infiltrativeSignature.Rmd <==
---
title: "Meta-programs explaining infiltrative vs nodular signature score"
author: "Massimo Andreatta & Santiago Carmona"
date: "`r format(Sys.Date(),'%e de %B, %Y')`"
output: 
  rmdformats::downcute:
    lightbox: true
    thumbnails: false
    self_contained: true
    gallery: true
    code_folding: show
  pkgdown:
    as_is: true
---

```{r, include=FALSE, fig.width=16, fig.height=12}
renv::restore()
library(Seurat)
library(ggplot2)
library(SignatuR)
library(UCell)
library(patchwork)
library(tidyr)
library(dplyr)
library(RColorBrewer)
library(viridis)
```


## Define config
```{r}
nprogs <- 7 #nr MPs
# if needed update MP variable names (e.g. WR) and interactions
```

## Set the paths

Read in source data
```{r}
seu <- readRDS("_aux/BCC_seurat_NMFprograms.rds")
```


```{r}
seu@misc$MP_signatures
seu@misc$MP_scale
seu@reductions[["NMFsignatures"]]
table(seu$Sample)
seu$batch <- factor(seu$orig.ident)
table(seu$batch)
```

Calculate the NI (nodular/infiltrative) score

```{r}
morphologySignatures <- read.csv("_aux/spatialSignatures.csv")
morphologySignatures <- as.list(morphologySignatures)
morphologySignatures <- lapply(morphologySignatures, function(x) x[x!=""])
seu <- AddModuleScore_UCell(seu, features = morphologySignatures, assay="RNA", ncores=4, name = "")
seu$ni_score <- seu$tumor_infiltrative-seu$tumor_nodular
hist(seu$ni_score)
```

```{r}
write.csv(x=seu@meta.data, file="cache/BCC_seurat_metadata.csv")
```


```{r}
sigs.score.data <- seu@meta.data[,c(grep("MP",names(seu@meta.data),value = T),"ni_score","patient_bcc","batch")]
colnames(sigs.score.data)[colnames(sigs.score.data) == "patient_bcc"] <- "Sample"
head(sigs.score.data)
```
```{r}
sigs.score.data$batch <- factor(sigs.score.data$batch, labels = c("BCC_Yerly","BCC_Ganier"))
```

Look at NI score per sample and batch/cohort

```{r}
ggplot(sigs.score.data,aes(x=ni_score))+geom_histogram()+facet_wrap(~Sample)+theme_bw()
```
```{r}
ggplot(sigs.score.data,aes(x=ni_score))+geom_histogram()+facet_wrap(~batch)+theme_bw()
```

Fit a linear model of MPs and samples as fixed-effect predictors and NI score as response variable

```{r}
myLm <- lm("ni_score ~ . -batch", data = sigs.score.data)
summary(myLm)
```



```{r}
library(jtools)
p <- plot_summs(myLm, ci_level = .999)
p
ggsave("plots/plot_coefs.pdf",p,width = 6, height = 4.5)
```

We can see that while globally some tumors are more nodular or more infiltrative than others and thus have some non-zero regression coefficients.
The largest regression estimate is however for MP7.WR, with a v alue of ~1 (i.e. holding other factors equal, an increase in MP7.WR Ucell score from 0 to 1 leads on average to an increase of NI score of 1 at the single-cell level)
Next, MP2 (MPPED1, LGR5) and MP4 (SFRP5) are negatively associated with NI score, and MP6 (FOSB, JUNB) is positively associated with NI score to a lesser degree.


Repeat lm without sample variables. Leads to very similar results.
```{r}
myLm_base <- lm("ni_score ~ . -Sample -batch", data = sigs.score.data)
summary(myLm_base)
p <- plot_summs(myLm_base, ci_level = .99)
p
ggsave("plots/plot_coefs_minModel.pdf",p,width = 6, height = 2)
ggsave("plots/plot_coefs_minModel.png",p,width = 6, height = 2)
```

And by excluding MP7?
Repeat lm without sample variables. Leads to very similar results.
```{r}
myLm_nn <- lm("ni_score ~ . -MP7.WR", data = sigs.score.data)
summary(myLm_nn)
p <- plot_summs(myLm_nn, ci_level = .99)
p

ggsave("plots/plot_coefs_noMP7.pdf",p,width = 6, height = 4.3)
```



Scatterplots of MP vs. NI score
```{r fig.height=6, fig.width=12}
library(ggrastr)
library(ggpubr)
pll <- lapply(names(seu@misc$MP_signatures), function(mp) {
  mp <- sym(mp)
  a <- ggplot(sigs.score.data, aes(x=!!mp, y=ni_score)) +
  geom_point(size=0.5) + scale_color_viridis(option="B") +
  stat_cor(p.digits=1, r.digits = 2, size=3) +
  geom_bin2d(bins=30) + 
  theme_minimal() + theme(aspect.ratio = 1)
  rasterize(a, layers='Point', dpi=200)
})

wrap_plots(pll, nrow=2)
ggsave("plots/MPs_vs_NI_scatter.pdf", height=7, width=13)
```

Split by sample
```{r fig.width=10}
a <- ggplot(sigs.score.data, aes(x=MP7.WR, y=ni_score)) +
  geom_point(size=0.5) + scale_color_viridis(option="B") +
    stat_cor(p.digits=1, r.digits = 2, size=3) +
     geom_bin2d(bins=30) +
  theme_bw() + theme(aspect.ratio = 1)
a + facet_wrap(~Sample, ncol = 6)

ggsave("plots/MP7_vs_NI_bysample.pdf", height=4, width=10)

```


Average correlation coefficient by patient
```{r}
corr <- lapply(names(seu@misc$MP_signatures), function(mp) {
  v <- c()
  for (s in unique(sigs.score.data$Sample)) {
    sub <- sigs.score.data[sigs.score.data$Sample == s,]
    v[s] <- cor(x=sub[,mp], y=sub[,"ni_score"], method="pearson")
  }
  v
}) 
names(corr) <- names(seu@misc$MP_signatures)
df <- as.data.frame(corr)

#sort by mean MP activity
avgs <- apply(df, 2, mean)
order <- sort(avgs, decreasing = T)

df$Sample <- rownames(df)
melt <- reshape2::melt(df)
colnames(melt) <- c("Sample","MP","Correlation")
melt$MP <- factor(melt$MP, levels=names(order))

cols <- c("red","forestgreen","blue","deeppink","darkcyan","orange","darkmagenta")
names(cols) <- names(seu@misc$MP_signatures)

a <- ggplot(melt, aes(y = Correlation, x = MP, fill=MP)) +
    scale_fill_manual(values=cols) +
    geom_hline(yintercept=0, linetype='dotted') +
    geom_boxplot(width=0.8, color="black", outlier.size = 2) +
    theme_classic2() +
    theme(axis.text.x = element_text(angle = 30, vjust = 1, hjust=1), axis.title.x = element_blank()) +
  ggtitle("Correlation with NI score by sample (N=11)")
a

ggsave("plots/MPs_vs_NI_correlation_by_sample.pdf", height=4.5, width=4.5)
```


```{r}
intervals <- levels(cut(sigs.score.data$ni_score, breaks = 3))

sigs.score.data$ni_class <- factor(cut(sigs.score.data$ni_score, breaks = 3),
                                   labels=c("nodular","intermediate","infiltrative"))
```

```{r}
sigs.score.data$NI.distrib <- "Dummy"

ggplot(sigs.score.data, aes(x=NI.distrib, y=ni_score)) + 
    geom_violin(trim=T) + theme_minimal()

ggsave("plots/NI.score_distribution.pdf", height=3.5, width=1.8)
```




Smooth gene signatures
```{r}
which.genes <- unique(unname(unlist(seu@misc$MP_signatures)))
seu <- SmoothKNN(obj=seu, signature.names = which.genes, assay="RNA", slot = "data")
```


Calculate MP specific specific genes
```{r}
MP.specific <- seu@misc$MP_signatures[c("MP3.ED","MP7.WR")]
MP.specific.genes <- unlist(MP.specific)
duplicated.genes <- MP.specific.genes[duplicated(unlist(MP.specific.genes))]
duplicated.genes

MP.specific <- lapply(MP.specific, function(x) x[!x %in% duplicated.genes])
names(MP.specific) <- paste0(names(MP.specific),".unique")
MP.specific
```

Plot expression of MP-specific genes:
```{r fig.width=12, fig.height=3.5}
genes <- c("TIMP1","CLDN4","CALML5","KRT6A","TYMP","CHI3L1","PHLDA2","IFITM3")

for (gene in genes) {
  p.scat.ED.WR <- ggplot(sigs.score.data, aes(x=MP3.ED, y=MP7.WR, colour=seu@assays$RNA_kNN@data[gene,])) +
    geom_point(alpha=0.5) + scale_color_viridis(option="F", direction=-1)  + coord_fixed(ratio = 1) + 
    theme_linedraw()
  p.scat.ED.WR$labels$colour <- gene
  a <- p.scat.ED.WR + facet_wrap(~ni_class)
  
  a <- rasterize(a, layers='Point', dpi=200)
  fname <- sprintf("plots/scatter_WDvsED_byNI_score_specificMPgenes_%s.pdf", gene)
  ggsave(fname, plot=a, width=9, height=2.8)
}
```


Calculate MP scores of specific genes
```{r}
MP.specific <- seu@misc$MP_signatures[c("MP3.ED","MP7.WR")]
MP.specific.genes <- unlist(MP.specific)
duplicated.genes <- MP.specific.genes[duplicated(unlist(MP.specific.genes))]
duplicated.genes

MP.specific <- lapply(MP.specific, function(x) x[!x %in% duplicated.genes])
names(MP.specific) <- paste0(names(MP.specific),".unique")
MP.specific
```

Explore gene set overlap between MPs and NI signatures

```{r}
library(UpSetR)

# example of list input (list of named vectors)
listInput <- append(seu@misc$MP_signatures, list(Nod=morphologySignatures$tumor_nodular,
                                                Infil=morphologySignatures$tumor_infiltrative))
upset(fromList(listInput), order.by = "freq",nsets = 20)
```


```{r}
library(ggvenn)
library(RColorBrewer)

venn1 <- ggvenn(listInput[c("MP3.ED","MP7.WR","Infil")], show_elements = F, label_sep = "\n", fill_color = brewer.pal(name="Set2",n=4), text_size = 4)
ggsave("plots/venn1.pdf")
venn2 <- ggvenn(listInput[c("MP3.ED","MP7.WR","Infil")], show_elements = T, label_sep = "\n", fill_color = brewer.pal(name="Set2",n=4), text_size = 0.6)
ggsave("plots/venn2.pdf")
venn3 <- ggvenn(listInput[c("MP3.ED","MP7.WR","Infil","Nod")], show_elements = F, label_sep = "\n", text_size = 4)
venn3
ggsave("plots/venn3.pdf")
wrap_plots(venn1,venn2,venn3)
```
