==> scRNA-seq/P6.1-wound_response_scRNAseq_BCC.Rmd <==
---
title: "Wound-response in CAF and cancer cells in scRNA-seq BCC patients retrospective cohort"
author: "Josep Garnica"
date: "`r format(Sys.Date(),'%e de %B, %Y')`"
output: 
  rmdformats::downcute:
    lightbox: true
    thumbnails: false
    self_contained: true
    gallery: true
    code_folding: show
  pkgdown:
    as_is: true
---

The goals of this script is to measure in the cell types of retrospective BCC patient cohort the:

  1. Wound-responding TME signatures (see script `P5-wounding_TME_signatures.Rmd`) for each corresponding celltype.
  2. Wound-related invasive (WRI), captured by MP7 metaprogram (see scripts from P1 to P3)

Finally, mean for each score is obtained per patient.

The score was computed on cells from different datasets: non-cancer, adjacent skin samples (“Normal adjacent”, N=6) obtained from GSE144236, Hedgehog inhibition (HHI)-therapy resistant BCC (N=7, samples from GSE123814), squamous cell carcinomas (N=5, samples from GSE144236), and BCC tumors (N=11, samples from GSE181907 and E-MTAB-13085) 



Load needed packages
```{r Load_packages, include=FALSE}
renv::restore()
library(Seurat)
require(scGate)
library(ggplot2)
library(UCell)
library(tidyr)
library(dplyr)
library(tibble)
library(ggprism)
library(data.table)

options(timeout=22222) # increase time out for downloading
```

# Load scGate models
## Fibroblasts, pericytes and smooth muscle cells (SMC)
Cancer-associated fibroblasts (CAF) were identified using the scGate method with specific fibroblast markers, excluding pericytes (PDGFA+, FBLN1+, FBLN2+, COL5A1+, LUM+, CFD+, RGS5-, NOTCH3-, MYOT-)


```{r Load_stromal_models}
pat <- "_scGate.tsv"
dir <- "_aux/"

#find files
mods.file <- list.files(dir, pattern = pat)

# read in files
mods <- lapply(mods.file, 
               function(f){
                 read.table(file.path(dir, f),
                            header = T)
               })
# adjust names
names(mods) <- gsub(pat, "", mods.file)
```

## Rest of cell type models

```{r scGate_models}
## HiTME models
allmodels <- scGate::get_scGateDB(branch = "1f650e5bfec2adb6ba121dcf90b6b0aef964c7ea",
                                  force_update = T)[["human"]][["HiTME"]]
hitmodels <- allmodels[-which(names(allmodels) == "Fibroblast")]

# change name to stromal as the true fibroblast signature contain different genes
hitmodels <- lapply(hitmodels,
                    function(x){
                      x %>% mutate(name = ifelse(name == "Fibroblast",
                                                 "Stromal", name))
                    })
```

Combine models
```{r combine_models}
joinmods <- c(mods, hitmodels)
```

# Load signatures
## wound-responding CAF

Load signatures previously produced in `P5-wounding_TME_signatures.Rmd`.
```{r Load_signatures}
# Load file with signatures
lys <- readRDS("cache/Wounding_signatures.rds")
```

## MP7 wound response on cancer cells
BCC cancer cells metaprograms were obtained in `P2-annotate_MPs.Rmd`
```{r load_MP7}
require(openxlsx)
mp <- read.xlsx("_aux/markers.MP.xlsx") %>% 
  as.list() %>% 
  lapply(., function(v){v[!is.na(v)]})
```



# Process retrospective datasets

## BCC tumors (GSE181907 and E-MTAB-13085)
Processing code of this datasets can be found:

* E-MTAB-13085: `1.1-Characterization_cell_populations.Rmd` for data from [Yerly et al., 2022](https://www.nature.com/articles/s41467-022-32670-w) -- "BCC_all_LY_clean_annotated.rds"

* GSE181907: `2.1-Characterization_cell_populations.Rmd` for data from [Ganier et al., 2023](https://www.pnas.org/doi/full/10.1073/pnas.2313326120#data-availability) -- "BCC_CG_all_annotated.rds"

### Load metadata
Load metadata obtained on `P3-MPs_vs_infiltrativeSignature.Rmd` script. It contains `ni_score` and `M7.WR` scoring for samples with annotated cancer cells (>100 annotated cancer cells).

```{r load_BCC_Scorings}
fpath <- "cache"
ni <- read.csv(file.path(fpath, "BCC_seurat_metadata.csv"))

# get the patients names with cancer cells and scored (11 patients)
keep_patients <- unique(ni$patient_bcc)
```


*Classify them as non-invasive or invasive:*
To identify non-invasive and invasive BCC tumors, samples from 11 patients having at least 100 cancer cells were classified using the following criteria. The average NI score by patient was calculated on their cancer cells. Based on the average NI score, patients were split into two groups of equal size: patients with an average NI score above the median were classified as invasive BCC, while those with an average NI score below the median were classified as Nodular. 


```{r}
nimed <- ni %>% 
  group_by(patient_bcc) %>% 
  summarize(mean_ni_score = mean(ni_score),
            Mean_MP7_score = mean(MP7.WR)) %>% 
  distinct() %>% 
  ungroup() %>% 
  mutate(category = ifelse(mean_ni_score < -0.02,
                           "Nodular",
                           "Infiltrative")
  )
```


### Load datasets
```{r load_BCC}
ly <- readRDS(file.path(fpath, "BCC_all_LY_clean_annotated.rds"))
cg <- readRDS(file.path(fpath, "BCC_CG_all_annotated.rds"))
```

split by patient
```{r split_BCC}
ly.spl <- SplitObject(ly, split.by = "patient_bcc")
cg.spl <- SplitObject(cg, split.by = "patient_bcc")

spl <- c(ly.spl, cg.spl)

rm(ly, cg, ly.spl, cg.spl)
gc()

# keep only patients with enough cancer cells (>100)
spl <- spl[keep_patients]
```


## Advanced BCC (GSE123814)
Hedgehog inhibition (HHI)-therapy resistant BCC
This dataset was download from [GEO](https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE123813) from [Yost et al., 2019](https://www.nature.com/articles/s41591-019-0522-3)

Processing of this datasets can be found at 5*.Rmd scripts. File loaded is produced in `5.1-Yost_characterization.Rmd`

### Load dataset
```{r load_Yost}
file <- "cache/BCC_Yost_all_annotated.rds"
yo <- readRDS(file)
```

### Run scGate
Classify cell types using previously loaded scGate models

```{r annotate_Yost}
yoli <- SplitObject(yo, "patient")
rm(yo); gc()
```


## Squamos cell carcinoma (SCC) and non-tumoral tissue [GSE144236]
This data was obtained from [GEO](https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE144236) and it is coming from this paper from [Ji et al., 2020](https://www.sciencedirect.com/science/article/pii/S0092867420306723?via%3Dihub) 

### Download dataset
```{r download_dir}
# create directory to download the data
ddir <- "cache/Ji"
dir.create(ddir, recursive = T)
```

#### Metadata
```{r Ji_metadata}
# download data
link <- "https://www.ncbi.nlm.nih.gov/geo/download/?acc=GSE144236&format=file&file=GSE144236%5Fpatient%5Fmetadata%5Fnew%2Etxt%2Egz"
dfile <- file.path(ddir, "GSE144236_patient_metadata_new.txt.gz")
download.file(url = link,
              destfile = dfile)

# load file
mdd <- read.table(dfile) %>% 
  # make unique id for each sample per patient
  mutate(sample = paste(patient, tum.norm, sep = "_"))
```

#### Matrix count

```{r Ji_counts}
# download data
link <- "https://www.ncbi.nlm.nih.gov/geo/download/?acc=GSE144236&format=file&file=GSE144236%5FcSCC%5Fcounts%2Etxt%2Egz"
dfile <- file.path(ddir, "GSE144236_cSCC_counts.txt.gz")
download.file(url = link,
              destfile = dfile)
# load file
co <- data.table::fread(dfile) %>% 
  as.matrix(rownames = "V1")

# remove first 2 rows that are actually metadata
co <- co[-c(1:2),] 
```

### Create Seurat object and Standarize gene names

```{r Create_seurat_standarize_Genes}
# create seurat object with previously obtained count matrix and metadata
ji <- CreateSeuratObject(counts = co,
                         meta.data = mdd,
                         min.features = 50,
                         min.cells = 0)
rm(co);gc()

# standarize gene symbols
require(STACAS)
# load gene symbols dataset
data(EnsemblGeneTable.Hs)
ji <- StandardizeGeneSymbols(ji,
                             slots = "counts",
                             EnsemblGeneTable = EnsemblGeneTable.Hs)
```

Normalize data
```{r Normalize and split-patient}
ji <- NormalizeData(ji)
```

### Standardize metadata
```{r standard_metadata_Ji}
ji$annotation <- ji$level2_celltype

ji$annotation[ji$annotation == "Tumor_KC"] <- "Cancer_cells"

ji$patient_bcc <- ji$sample
```



split object
```{r split_JiDataset}

jili <- SplitObject(ji, "sample")

```


## Join datasets
```{r join_list_Datasets}
bcca <- c(yoli, jili, spl)
rm(spl, yoli, jili); gc()

saveRDS(bcca, "cache/retrospective_all_list.rds")
```

# Classify cell types

```{r classify_scGate}
done <- c()
for(a in names(bcca)[!names(bcca) %in% done]){
  # we can do just one scGate
  cat("####################### Doing scGate for", a, "#######################\n")
  if("scGate_multi" %in% names(bcca[[a]]@meta.data)){
    bcca[[a]]$sGate_multi0 <- bcca[[a]]$scGate_multi
  }
  
  bcca[[a]] <- scGate::scGate(bcca[[a]],
                              model = joinmods,
                              ncores = 8)
  
  done <- c(done, a)
  cat("--------------- ", length(done), " / ", length(bcca), " -------------\n\n")
}
```



# Signature scoring
Since Ucell scores signatures cell-wise we can score our signatures on the whole dataset all at once and then filter cells.

We are interested in measuring previously discovered programs (MPs, see `P1-NMF.Rmd`) in cancer cells and TME signatures for CAF/Fibroblasts, MoMac, DC, pericytes, which show most DE genes in wound vs non-wound samples.

```{r join_signatures}
# join signatures
tme_celltypes <- ct <- c("Fibroblast", "MoMac", "panDC","Pericyte")
signatures <- append(lys[tme_celltypes], mp)
```

## Filter dataset for these cell types
```{r filter_Dataset}
bcca.filt <- bplapply(bcca,
                      BPPARAM = bparam,
                      function(s){
                        # filter for 
                        if("annotation")
                          # remove doublets
                          s <- s[,s$annotation != "Doublets"]
                        # obtain cell types
                          sub <- s[,s$scGate_multi %in% tme_celltypes &
                                     s$annotation == "Cancer_cells"]
                        return(sub)
                      })
# merge
bcca.mer <- merge(bcca.filt[[1]], bcca.filt[-1])
rm(bcca.filt, bcca); gc()
```


## Scoring UCell

```{r score_ucell}
bcca.mer <- AddModuleScore_UCell(bcca.mer,
                                 features = signatures,
                                 ncores = 8,
                                 name = "")
```


## Wound-responding TME
### dataframe
```{r df_TME_Scoring}
df <- bcca.mer@meta.data[,grepl("ound|scGate_multi$|patient_bcc", names(bcca.mer@meta.data))] %>% 
  group_by(patient_bcc, scGate_multi) %>% 
  summarize(across(matches("ound"), mean, .names = "{.col}"),
            n = n()) %>% 
  # filter cells below 10
  filter(n > 9)   %>% 
  dplyr::select(-n) %>% 
  tidyr::pivot_longer(-c(patient_bcc, scGate_multi),
                      names_to = "mm",
                      values_to = "Score") %>% 
  tidyr::separate(mm, sep = "[.]", into = c("celltype", "signature")) %>% 
  filter(scGate_multi == celltype) %>% 
  tidyr::pivot_wider(names_from = "signature",
                     values_from = "Score") %>% 
  mutate(Score = Wound - Unwound) %>% 
  left_join(., nimed[,c("patient_bcc", "category")], by = "patient_bcc") %>% 
  mutate(category = ifelse(is.na(category),
                           "Normal adjacent",
                           as.character(category)),
         category = factor(category,
                           levels = c("Normal adjacent",
                                      "Nodular",
                                      "Infiltrative")
         ),
         celltype = factor(celltype,
                           levels = c("Fibroblast", "MoMac",
                                      "panDC", "Perycyte"),
                           labels = c("Fibroblast/CAF", "MoMac",
                                      "panDC", "Pericyte"))
  )
```

### Plot
```{r plot_wr_TME}
require(ggpubr)
lycols <- c("Normal adjacent" = "oldlace",
            "Nodular" = "#FFFF6C",
            "Infiltrative" = "#F44400")

comps <- list(
              c("Normal adjacent", "Nodular"),
              c("Infiltrative", "Nodular"))

set.seed(22)
pl <- df %>% 
  ggplot(aes(category, Score, fill = category)) +
  geom_boxplot(outlier.colour = NA,
               show.legend = F,
               width = 0.6) +
  geom_jitter(width = 0.1,
              size = 2,
              shape = 21,
              fill = "white",
              show.legend = F) +
  stat_compare_means(
    comparisons = comps,
    method = "wilcox.test", 
    label = "p.format",
    p.adjust.method = "BH",
    size = 3.5
  ) +
scale_fill_manual(values = lycols) +
  scale_x_discrete(labels = gsub(" ", "\n", levels(df$category))) +
  labs(y = "Mean wound-responding score",
       x = "",) +
  facet_wrap(~celltype,
             ncol = 2, 
             scales = "free_y") +
  theme_classic2() +
  theme(axis.text.x = element_text(angle = 30, hjust = 1, vjust = 1))
```



## MPs scoring
### Summarize dataframe
```{r}
df <- bcca.mer@meta.data[,grepl("MP|patient_bcc", names(bcca.mer@meta.data))] %>% 
  group_by(patient_bcc) %>% 
  summarize(across(matches("MP"), mean, .names = "{.col}")) %>% 
  left_join(., nimed[,c("patient_bcc", "category")], by = "patient_bcc") %>% 
  mutate(category = ifelse(is.na(category),
                           ifelse(grepl("_Tum", patient_bcc),
                                  "SCC", "Advanced BCC"),
                           as.character(category)),
         category = factor(category,
                           levels = c("Nodular","Infiltrative",
                                      "Advanced BCC", "SCC")
         )
  ) %>% 
  tidyr::pivot_longer(-c(patient_bcc, category),
                      names_to = "Metaprogram",
                      values_to = "Score")
  
```

### Plot
```{r}
require(ggpubr)
lycols <- c(
            "Nodular" = "#FFFF6C",
            "Infiltrative" = "#F44400",
            "Advanced BCC" = "red4",
            "SCC" = "violetred")

comps <- list(c("SCC", "Advanced BCC"),
              c("Advanced BCC", "Infiltrative"),
              c("Infiltrative", "Nodular"))

set.seed(22)
pl <- df %>% 
  ggplot(aes(category, Score, fill = category)) +
  geom_boxplot(outlier.colour = NA,
               show.legend = F,
               width = 0.6) +
  geom_jitter(width = 0.1,
              size = 2.5,
              shape = 21,
              fill = "white",
              show.legend = F) +
  stat_compare_means(
    comparisons = comps,
    method = "wilcox.test", 
    label = "p.format",
    p.adjust.method = "BH",
    size = 3.5
  ) +
scale_fill_manual(values = lycols) +
  scale_x_discrete(labels = gsub(" ", "\n", levels(df$category))) +
  labs(y = "Mean score in cancer cells",
       x = "",) +
  facet_wrap(~Metaprogram,
             ncol = 4, 
             scales = "free_y") +
  theme_classic2() +
  theme(axis.text.x = element_text(angle = 30, hjust = 1, vjust = 1))

```
